Optimize attendance query by filtering dates
