+++ pkg/web/attendance_handlers.go
@@ -25,6 +25,10 @@ func (s Server) OvertimeReport() http.Handler {
 		forAnotherUser := r.FormValue("userscope") == "user-foreign-radio"
 		searchUser := r.FormValue("username")
 
+		year := parseIntOrDefault(r.FormValue("year"), time.Now().Year())
+		month := parseIntOrDefault(r.FormValue("month"), int(time.Now().Month()))
+		fte := parseFloatOrDefault(r.FormValue("ftepercentage"), 100)
+
 		var employee *odoo.Employee
 		if forAnotherUser {
 			e, err := s.odoo.SearchEmployee(searchUser, session.ID)
@@ -34,6 +38,10 @@ func (s Server) OvertimeReport() http.Handler {
 			}
 			if e == nil {
 				view.ShowError(w, fmt.Errorf("no user matching '%s' found", searchUser))
+				return
+			}
+			if !e.AttendanceAccess {
+				view.ShowError(w, fmt.Errorf("you don't have access to read attendances of '%s'", e.Name))
 				return
 			}
 			employee = e
@@ -46,7 +54,9 @@ func (s Server) OvertimeReport() http.Handler {
 			employee = e
 		}
 
-		attendances, err := s.odoo.ReadAllAttendances(session.ID, employee.ID)
+		begin := odoo.Date(time.Date(year, time.Month(month), 1, 0, 0, 0, 0, time.UTC))
+		end := odoo.Date(begin.ToTime().AddDate(0, 1, 0))
+		attendances, err := s.odoo.FetchAttendancesBetweenDates(session.ID, employee.ID, begin, end)
 		if err != nil {
 			view.ShowError(w, err)
 			return
@@ -57,10 +67,6 @@ func (s Server) OvertimeReport() http.Handler {
 			view.ShowError(w, err)
 			return
 		}
-
-		year := parseIntOrDefault(r.FormValue("year"), time.Now().Year())
-		month := parseIntOrDefault(r.FormValue("month"), int(time.Now().Month()))
-		fte := parseFloatOrDefault(r.FormValue("ftepercentage"), 100)
 
 		reporter := timesheet.NewReporter(attendances, leaves, employee).SetFteRatio(fte/100).SetMonth(year, month)
 		report := reporter.CalculateReport()
